Byte is received successfully, but xTimerStop is delaying
This commit is because I can capture the byte received from UART one
time and then I want to stop timer to avoid overriding of current_byte,
but this task of stop/delete is delaying and the current_byte is being
override by 1's after the stop bit happens. Please fix this.