PWM config implemented and byte received from virtual UART passed to fcn
This commit is adding all necessary config for the 4 channels of PWM,
but PWM output is not executing correctly and now the current byte
received is being transformed wrongly as it is going to receive 8 bits
with the correct number in decimal for percentage of the duty cycle.
Missing to fix this is necessary, next step is to generate frame TX with
FreeRTOS in the meantime.